// src/app/api/model/pensionStatusEntryDTO.ts
/**
 * CTS-BE
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateOnly } from './dateOnly';


export interface PensionStatusEntryDTO { 
    dataSource?: { [key: string]: any; } | null;
    statusFlag: number;
    statusWef: DateOnly;
    ppoId: number;
}